[MCUX-78503][MCXE31]Fix MSG issues for clock driver.
- Fixed some CERT INT30-C issue.

+++ MCXE/MCXE31B/system_MCXE31B.c
@@ -100,9 +100,9 @@ void SystemCoreClockUpdate(void)
                 {
                     freq   = CLK_XTAL_OSC_CLK / 1000U / div;
                     temp   = (PLL->PLLDV & PLL_PLLDV_MFI_MASK) >> PLL_PLLDV_MFI_SHIFT;
-                    temp64 = (uint64_t)temp * 18432U + (PLL->PLLFD & PLL_PLLFD_MFN_MASK);
+                    temp64 = (uint64_t)temp * 18432U + (uint64_t)(PLL->PLLFD & PLL_PLLFD_MFN_MASK);
                     temp64 = temp64 * 1000U / 18432U;
-                    freq   = (uint32_t)(freq * temp64);
+                    freq   = (uint32_t)((uint64_t)freq * temp64);
                 }
                 else
                 {
